+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/KeystoreProvider.kt
@@ -23,11 +23,14 @@ import at.asitplus.wallet.lib.agent.KeyMaterial
 import at.asitplus.wallet.lib.agent.KeyWithSelfSignedCert
 import data.storage.DataStoreService
 import io.github.aakira.napier.Napier
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.IO
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
+import kotlinx.coroutines.withContext
 
 
 private const val CERT_STORAGE_KEY = "MB64_CERT_SELF_SIGNED"
@@ -35,25 +38,17 @@ private const val CERT_STORAGE_KEY = "MB64_CERT_SELF_SIGNED"
 open class KeystoreService(
     private val dataStoreService: DataStoreService
 ) {
-    private val sMut = Mutex()
+    private val dispatcher = Dispatchers.IO.limitedParallelism(1)
 
     @Throws(Throwable::class)
-    open suspend fun getSigner(): KeyMaterial {
-        var signer: KeyMaterial? = null
-        Napier.d("getSigner")
-        sMut.withLock {
-            if (signer == null)
-                signer = catchingUnwrapped { initSigner() }
-                    .getOrElse { FallBackKeyMaterial(it) }
-        }
-        return signer!!
-    }
+    open suspend fun getSigner(): KeyMaterial = catchingUnwrapped { initSigner() }
+        .getOrElse { FallBackKeyMaterial(it) }
 
     @Throws(Throwable::class)
-    private suspend fun initSigner(): KeyWithSelfSignedCert {
+    private suspend fun initSigner(alias: String): KeyWithSelfSignedCert = withContext(dispatcher) {
         PlatformSigningProvider.let { provider ->
-            val forKey = provider.getSignerForKey(Configuration.KS_ALIAS).getOrElse {
-                provider.createSigningKey(alias = Configuration.KS_ALIAS) {
+            val forKey = provider.getSignerForKey(alias).getOrElse {
+                provider.createSigningKey(alias = alias) {
                     ec {
                         curve = ECCurve.SECP_256_R_1
                         purposes {
@@ -72,8 +67,34 @@ open class KeystoreService(
                     }
                 }.getOrThrow()
             }
-            return KeyWithPersistentSelfSignedCert(forKey, forKey.publicKey.didEncoded, 30)
+            KeyWithPersistentSelfSignedCert(forKey, forKey.publicKey.didEncoded, 30)
         }
+    }
+
+    @Throws(Throwable::class)
+    private suspend fun initSigner(): KeyWithSelfSignedCert = initSigner(Configuration.KS_ALIAS)
+
+    open suspend fun testSigner(): Boolean = withContext(dispatcher) {
+        runCatching {
+            PlatformSigningProvider.let { provider ->
+                provider.deleteSigningKey(Configuration.KS_CAPABILITY_ALIAS)
+                provider.createSigningKey(Configuration.KS_CAPABILITY_ALIAS)
+            }
+        }.isSuccess
+    }
+
+    suspend fun testAttestation() = withContext(dispatcher) {
+        runCatching {
+            PlatformSigningProvider.deleteSigningKey(Configuration.KS_CAPABILITY_ALIAS)
+            PlatformSigningProvider.createSigningKey(Configuration.KS_CAPABILITY_ALIAS) {
+                ec {}
+                hardware {
+                    attestation {
+                        this.challenge = "CHALLENGE".encodeToByteArray()
+                    }
+                }
+            }.getOrThrow()
+        }.isSuccess
     }
 
     inner class KeyWithPersistentSelfSignedCert(