+++ shared/src/commonMain/kotlin/ui/viewmodels/InitializationViewModel.kt
@@ -1,0 +1,8 @@
+package ui.viewmodels
+
+import androidx.lifecycle.ViewModel
+import at.asitplus.wallet.app.common.WalletMain
+
+class InitializationViewModel(
+    val walletMain: WalletMain,
+) : ViewModel()