+++ shared/src/commonMain/kotlin/ui/views/SettingsView.kt
@@ -88,6 +88,7 @@ fun SettingsView(
     onClickDataProtectionPolicy: (() -> Unit)?,
     onClickLicenses: (() -> Unit)?,
     koinScope: Scope,
+    onReset: () -> Unit,
     settingsViewModel: SettingsViewModel = koinViewModel(scope = koinScope),
 ) {
     var isLoading by rememberSaveable {
@@ -101,6 +102,7 @@ fun SettingsView(
                 showResetAlert = false
                 isLoading = true
                 settingsViewModel.onClickResetApp {
+                    onReset()
                     isLoading = false
                     if (it != null) {
                         settingsViewModel.showGlobalSnackbar {