+++ shared/src/iosMain/kotlin/main.ios.kt
@@ -6,19 +6,63 @@ import at.asitplus.KmmResult
 import at.asitplus.dcapi.request.DCAPIRequest
 import at.asitplus.iso.EncryptionParameters
 import at.asitplus.wallet.app.common.BuildContext
+import at.asitplus.wallet.app.common.CapabilitiesService
 import at.asitplus.wallet.app.common.KeystoreService
 import at.asitplus.wallet.app.common.PlatformAdapter
+import at.asitplus.wallet.app.common.RealCapabilitiesService
+import at.asitplus.wallet.app.common.SESSION_NAME
 import at.asitplus.wallet.app.common.WalletDependencyProvider
 import at.asitplus.wallet.app.common.dcapi.data.export.CredentialRegistry
+import at.asitplus.wallet.app.common.di.appModule
 import data.storage.RealDataStoreService
 import data.storage.createDataStore
 import io.github.aakira.napier.Napier
-import kotlinx.cinterop.*
+import kotlinx.cinterop.BetaInteropApi
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.cinterop.ObjCObjectVar
+import kotlinx.cinterop.addressOf
+import kotlinx.cinterop.alloc
+import kotlinx.cinterop.memScoped
+import kotlinx.cinterop.ptr
+import kotlinx.cinterop.usePinned
+import kotlinx.cinterop.value
 import kotlinx.coroutines.CoroutineScope
+import org.koin.core.module.dsl.scopedOf
+import org.koin.core.qualifier.named
+import org.koin.dsl.binds
+import org.koin.dsl.module
 import org.multipaz.compose.prompt.PromptDialogs
 import org.multipaz.prompt.IosPromptModel
-import platform.Foundation.*
-import platform.UIKit.*
+import platform.AVFoundation.AVAuthorizationStatusAuthorized
+import platform.AVFoundation.AVAuthorizationStatusDenied
+import platform.AVFoundation.AVAuthorizationStatusNotDetermined
+import platform.AVFoundation.AVAuthorizationStatusRestricted
+import platform.AVFoundation.AVCaptureDevice
+import platform.AVFoundation.AVMediaTypeVideo
+import platform.AVFoundation.authorizationStatusForMediaType
+import platform.Foundation.NSData
+import platform.Foundation.NSDocumentDirectory
+import platform.Foundation.NSError
+import platform.Foundation.NSFileHandle
+import platform.Foundation.NSFileManager
+import platform.Foundation.NSSet
+import platform.Foundation.NSString
+import platform.Foundation.NSURL
+import platform.Foundation.NSUTF8StringEncoding
+import platform.Foundation.NSUserDomainMask
+import platform.Foundation.anyObject
+import platform.Foundation.closeFile
+import platform.Foundation.create
+import platform.Foundation.fileHandleForWritingAtPath
+import platform.Foundation.seekToEndOfFile
+import platform.Foundation.stringWithContentsOfFile
+import platform.Foundation.writeData
+import platform.UIKit.UIActivityViewController
+import platform.UIKit.UIApplication
+import platform.UIKit.UIApplicationOpenSettingsURLString
+import platform.UIKit.UIViewController
+import platform.UIKit.UIWindow
+import platform.UIKit.UIWindowScene
 import platform.darwin.dispatch_async
 import platform.darwin.dispatch_get_main_queue
 import ui.theme.darkScheme
@@ -42,18 +86,23 @@ fun MainViewController(
     val dataStoreService = RealDataStoreService(createDataStore(), iosPlatformAdapter)
     val keystoreService = KeystoreService(dataStoreService)
     val promptModel = IosPromptModel()
+    val walletDependencyProvider = WalletDependencyProvider(
+        keystoreService,
+        dataStoreService,
+        iosPlatformAdapter,
+        buildContext = buildContext,
+        promptModel = promptModel
+    )
+    val capabilitiesModule = module {
+        scope(named(SESSION_NAME)) {
+            scopedOf(::RealCapabilitiesService) binds arrayOf(CapabilitiesService::class)
+        }
+    }
+    val module = appModule(walletDependencyProvider, capabilitiesModule)
 
     return ComposeUIViewController {
         PromptDialogs(promptModel)
-        App(
-            WalletDependencyProvider(
-                keystoreService,
-                dataStoreService,
-                iosPlatformAdapter,
-                buildContext = buildContext,
-                promptModel = promptModel
-            )
-        )
+        App(module)
     }
 }
 
@@ -220,6 +269,10 @@ class IosPlatformAdapter(
         //TODO("Not yet implemented")
     }
 
+    override fun openDeviceSettings() {
+        openUrl(UIApplicationOpenSettingsURLString)
+    }
+
     fun getBaseUrl(): NSURL? {
         val urls = NSFileManager.defaultManager.URLsForDirectory(
             directory = NSDocumentDirectory,
@@ -228,9 +281,21 @@ class IosPlatformAdapter(
         return urls.first() as? NSURL
     }
 
+    override fun getCameraPermission(): Boolean? {
+        val status = AVCaptureDevice.authorizationStatusForMediaType(AVMediaTypeVideo)
+
+        return when (status) {
+            AVAuthorizationStatusAuthorized -> true
+            AVAuthorizationStatusNotDetermined -> null
+            AVAuthorizationStatusDenied -> false
+            AVAuthorizationStatusRestricted -> false
+            else -> null
+        }
+    }
+
 }
 
-@OptIn(ExperimentalForeignApi::class)
+@OptIn(ExperimentalForeignApi::class, BetaInteropApi::class)
 fun ByteArray.toNSData(): NSData = memScoped {
     this@toNSData.usePinned { pinned ->
         NSData.create(