+++ CHANGELOG.md
@@ -12,6 +12,9 @@
  * Update to multipaz 0.94.0
  * Add ble and nfc capability checks for proximity (enabled, selected)
  * Move transfer options selection from general settings view to proximity holder and verifier route
+ * Add CapabilitiesService, CapabilitiesView to hint wrong settings
+ * Add InitializationView
+ * Add AuthCheckKit as new dependency
 
 # Release 5.7.3
  * Update to VC-K 5.8.0, fixing optional attributes during presentation